Done: finished code main logic

+++ main.py
@@ -2,24 +2,27 @@
 from datetime import datetime, timedelta ## https://www.geeksforgeeks.org/python/python-datetime-timedelta-function/
 import openmeteo_requests ## https://open-meteo.com/en/docs
 import os
-import csv
+from zoneinfo import ZoneInfo
 import pandas as pd
 import requests_cache
 from retry_requests import retry
 
 def thirtyDayCount():
     file_path = os.path.join("30daysReg")
-    for file in os.listdir(file_path):
-        thirtyArr = []
-        fileTarget = os.path.basename(file).split('-')
-        ano, mes, dia = fileTarget[1].split("M")[0], fileTarget[2].split("D")[0], fileTarget[2].split("D")[1]
-        thirtyArr.append((int(ano), int(mes), int(dia), file))
+    if not os.path.exists(file_path):
+        os.makedirs(folder, exist_ok=True)
+    else:
+        for file in os.listdir(file_path):
+            thirtyArr = []
+            fileTarget = os.path.basename(file).split('-')
+            ano, mes, dia = fileTarget[1].split("M")[0], fileTarget[2].split("D")[0], fileTarget[2].split("D")[1]
+            thirtyArr.append((int(ano), int(mes), int(dia), file))
 
-    #no loop because the code will run every day( in thesis )
-    if len(thirtyArr) > 30:
-        menor_data = min(thirtyArr)
-        older = menor_data[3]
-        os.remove(os.path.join(file_path, older))
+        #no loop because the code will run every day( in thesis )
+        if len(thirtyArr) > 30:
+            menor_data = min(thirtyArr)
+            older = menor_data[3]
+            os.remove(os.path.join(file_path, older))
 
 def prev24count(ano, mes, dia, hora):
     prev_hours = []
@@ -41,7 +44,6 @@ def prev24count(ano, mes, dia, hora):
 
 
 def hoursVarianceVarReads(ano, mes, dia, hora):
-
     metrics = [
         "temperature_2m",
         "rain",
@@ -60,16 +62,16 @@ def hoursVarianceVarReads(ano, mes, dia, hora):
         path = f"30daysReg/A-{a}M-{m}D{d}-H{h}.csv"
         if os.path.exists(path):
 
-            df = pd.read_csv(path)
+            df = pd.read_csv(path, parse_dates=["date"])
 
-            target_time = f"{a}-{m:02d}-{d:02d} {h:02d}:00:00-07:00"
+            df["date"] = pd.to_datetime(df["date"], utc=True).dt.tz_convert(ZoneInfo("America/Los_Angeles"))
+            target_time = pd.Timestamp(year=a, month=m, day=d, hour=h, tzinfo=ZoneInfo("America/Los_Angeles"))
             match = df.loc[df["date"] == target_time]
 
             if match.empty:
                 continue
 
             for metric in metrics:
-
                 value = float(match.iloc[0][metric])
                 metric_values[metric].append(value)
 
@@ -96,13 +98,18 @@ openmeteo = openmeteo_requests.Client(session = retry_session)
 # Make sure all required weather variables are listed here
 # The order of variables in hourly or daily is important to assign them correctly below
 url = "https://api.open-meteo.com/v1/forecast"
+
+metrics = ["temperature_2m", "rain", "precipitation_probability", "relative_humidity_2m", "wind_speed_10m"]
+models = "gfs_seamless"
+timezone = "America/Los_Angeles"
+
 params = {
 	"latitude": 47.6062,
 	"longitude": -122.3321,
-	"hourly": ["temperature_2m", "rain", "precipitation_probability", "relative_humidity_2m", "wind_speed_10m"],
-	"models": "gfs_seamless",
-	"current": ["temperature_2m", "rain", "precipitation_probability", "relative_humidity_2m", "wind_speed_10m"],
-	"timezone": "America/Los_Angeles",
+	"hourly": metrics,
+	"models": models,
+	"current": metrics,
+	"timezone": timezone,
 	"forecast_days": 3,
 }
 responses = openmeteo.weather_api(url, params = params)
@@ -137,14 +144,14 @@ hourly_precipitation_probability = hourly.Variables(2).ValuesAsNumpy()
 hourly_relative_humidity_2m = hourly.Variables(3).ValuesAsNumpy()
 hourly_wind_speed_10m = hourly.Variables(4).ValuesAsNumpy()
 
-hourly_data = {
+hourly_data = pd.DataFrame({
 	"date": pd.date_range(
 		start = pd.to_datetime(hourly.Time(), unit = "s", utc = True),
 		end =  pd.to_datetime(hourly.TimeEnd(), unit = "s", utc = True),
 		freq = pd.Timedelta(seconds = hourly.Interval()),
 		inclusive = "left"
 	).tz_convert(response.Timezone().decode())
-}
+})
 
 hourly_data["temperature_2m"] = hourly_temperature_2m
 hourly_data["rain"] = hourly_rain
@@ -152,34 +159,20 @@ hourly_data["precipitation_probability"] = hourly_precipitation_probability
 hourly_data["relative_humidity_2m"] = hourly_relative_humidity_2m
 hourly_data["wind_speed_10m"] = hourly_wind_speed_10m
 
-
-
-
-current_time = (
-    pd.to_datetime(current.Time(), unit="s", utc=True)
-    .tz_convert(response.Timezone().decode())
-    .strftime("%Y-%m-%d %H")
-)
-current_time = current_time.rsplit()
-ano = int(current_time[0][0:4])
-mes = int(current_time[0][5:7])
-dia = int(current_time[0][8:10])
-hora = int(current_time[1])
-file_path = os.path.join("30daysVariance", "variance.csv")
-
-OntimeRegister = pd.read_csv(file_path)
-
+current_time_dt = datetime.fromtimestamp(current.Time(),tz=ZoneInfo("America/Los_Angeles")).replace(minute=0, second=0, microsecond=0)
+ano = current_time_dt.year
+mes = current_time_dt.month
+dia = current_time_dt.day
+hora = current_time_dt.hour
 varVarianceList = hoursVarianceVarReads(ano, mes, dia, hora)[1]
+
+
 for metric, value in varVarianceList.items():
-    OntimeRegister.loc[dia, f"{metric}_variance"] = round(abs(value -  locals()[f"current_{metric}"]),2)
-
-OntimeRegister.to_csv(file_path, index=False)
-
-
-###################
-hourly_dataframe = pd.DataFrame(hourly_data)
+    metric_variance_col = f"{metric}_variance"
+    hourly_data[metric_variance_col] = 0.0
+    variance_value = round(abs(value -  locals()[f"current_{metric}"]),2)
+    hourly_data.loc[dia, metric_variance_col] = variance_value
 
 for folder in ["30daysReg"]:
     thirtyDayCount()
-    os.makedirs(folder, exist_ok=True)
-    hourly_dataframe.to_csv(f"{folder}/A-{ano}M-{mes}D{dia}-H{hora}.csv", index=False)
+    hourly_data.to_csv(f"30daysReg/A-{ano}M-{mes}D{dia}-H{hora}.csv", index=False)
